microservices/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
azure-storage==0.36.0
azure-functions==1.2.0
dataclasses-json==0.4.2
jsonpickle==1.3
grpcio-tools==1.28.1
grpcio==1.28.1
phonenumbers==8.12.1
pymongo==3.10.1